+++ R/app_server.R
@@ -396,11 +396,11 @@ app_server <- function(input, output, session) {
         if (input$testing) {cat("should not be here\n")}
       } else {
         #   ###################################### #
-        # file or object (spatial data.frame) was provided as shapefile param in ejamapp()
-        if (input$testing) {cat("trying to read shapefile specified in ejamapp() call \n")}
-        shp <- try( shapefile_from_any(xshp, cleanit = FALSE, silentinteractive = TRUE), silent = TRUE)
-        if (inherits(shp, "try-error")) {
-          cat("failed to read shapefile specified in ejamapp() call \n")
+        # if the file or object (spatial data.frame) was provided as shapefile param in ejamapp()
+        if (input$testing) {cat("trying to read shapefile parameter\n")}
+        shp <- try( shapefile_from_any(xshp, cleanit = FALSE, silentinteractive=TRUE), silent = !testing)
+        if (is.null(shp) || inherits(shp, "try-error")) {
+          cat("shapefile_from_any() cannot read specified shp parameter \n")
           req(FALSE, cancelOutput = TRUE)
         }
         ## if user provided ejamapp(shapefile=xyz) but did not set these also, they will not see their upload ready to run:
@@ -433,14 +433,16 @@ app_server <- function(input, output, session) {
     #   ###################################### #
     # do the rest whether it was uploaded or came via ejamapp()
 
-    # if shp contains point features, present message in app
-    ## this case is not caught by shapefile_from_any currently - but could use shapefix somehow? ***
-    if (any(sf::st_geometry_type(shp) == "POINT")) {
-      shp <- NULL
-      disable_buttons[['SHP']] <- TRUE
-      msg <- "Shape file must be of polygon geometry."
-      cat(msg, "\n")
-      shiny::validate(msg)
+    if (!is.null(shp)) {
+      # if shp contains point features, present message in app
+      ## this case is not caught by shapefile_from_any currently - but could use shapefix somehow? ***
+      if (any(sf::st_geometry_type(shp) == "POINT")) {
+        shp <- NULL
+        disable_buttons[['SHP']] <- TRUE
+        msg <- "Shape file must be of polygon geometry."
+        cat(msg, "\n")
+        shiny::validate(msg)
+      }
     }
     # note that shapefile_from_any() returns some info in attributes of the returned object, like error messages and counts of valid points
     if (!is.null(attr(shp, "validate_errmsg")))            {shiny::validate( attr(shp, "validate_errmsg") )}
@@ -1711,7 +1713,7 @@ app_server <- function(input, output, session) {
       valid <- !is.na(data_uploaded()$lon) & !is.na(data_uploaded()$lat)
       valid_pts_count <- NROW(data_uploaded()[valid, ])
       if (valid_pts_count > max_pts_map) { # would have already been stopped probably
-       ## Max allowed points FOR MAP was exceeded!
+        ## Max allowed points FOR MAP was exceeded!
         if (valid_pts_count > max_pts_run) {
           ## Max allowed points FOR ANALYSIS was exceeded also, not just cap on mapping!
           msg <- paste0('Too many valid points (',
@@ -2274,6 +2276,7 @@ app_server <- function(input, output, session) {
         dplyr::select(-any_of(c('valid', 'invalid_msg'))) %>%
         sf::st_zm() %>% sf::as_Spatial() # st_zm() was already done? ***
 
+      # d_uploads is an object of class "SpatialPolygonsDataFrame" not "sf" and "data.frame" like data_uploaded() here is
       leaflet::leafletProxy(mapId = 'an_leaf_map', session) %>%
         map_shapes_leaflet_proxy(shapes = d_uploads, popup = popup_from_df(d_uploads %>% sf::st_drop_geometry()))