+++ tests/testthat/test-url_ejamapi.R
@@ -1,4 +1,13 @@
 
+# see ?ejamapi() and test-ejamapi.R for testing the wrapper that is an interface to
+#   the "report" and "data" endpoints, using GET and POST, for the live EJAM API that is hosted on a server.
+
+# see ?url_ejamapi() and test-url_ejamapi.R for testing the utility that provides a URL
+#   (that handles the "report" endpoint, and is designed to work in the live EJAM API that is hosted on a server).
+
+# see ?ejamapi_local() and test-ejamapi_local.R for testing the DRAFT-ONLY API defined in the EJAM package, run locally.
+
+
 # test-url_ejamapi.R
 
 # tests  ####